refactor: page 에서 PostsManager / PostTable 분리 및 모달 추가 분리
